Do not leave stdout closed in daemon mode
When sensord daemonizes, it closes stderr and stdout files. It then
assigns stderr to be a log file. stdout is kept closed, however sensord
still tries to output to it with regular `printf` statements.

This messes communication with /dev/i2c-1 device, which is being
assigned file descriptor "1", typically used by stdout.

Assign stdout to log file, so that prints are routed to the right place.

Fixes #24.