main: add option -i (inetd mode)
This allows launching sensord with systemd socket activation; this
switches sensord from a variod client to a server which listens for an
incoming connection from variod.  This is a major code simplification,
and will allow eliminating lots of code from sensord (once the old
mode can be removed).

For now, the old mode remains, and the inetd mode is just a new option.

Closes https://github.com/Openvario/sensord/issues/42